+++ R/data_read.R
@@ -15,15 +15,16 @@
 #' for SAS data files.
 #' @param encoding The character encoding used for the file. Usually not needed.
 #' @param convert_factors If `TRUE` (default), numeric variables, where all
-#' values have a value label, are assumed to be categorical and converted
-#' into factors. If `FALSE`, no variable types are guessed and no conversion
-#' of numeric variables into factors will be performed. See also section
-#' 'Differences to other packages'. For `data_write()`, this argument only
-#' applies to the text (e.g. `.txt` or `.csv`) or spreadsheet file formats (like
-#' `.xlsx`). Converting to factors might be useful for these formats because
-#' labelled numeric variables are then converted into factors and exported as
-#' character columns - else, value labels would be lost and only numeric values
-#' are written to the file.
+#' values have a value label, are assumed to be categorical and converted into
+#' factors. If `FALSE`, no variable types are guessed and no conversion of
+#' numeric variables into factors will be performed. For `data_read()`, this
+#' argument only applies to file types with *labelled data*, e.g. files from
+#' SPSS, SAS or Stata. See also section 'Differences to other packages'. For
+#' `data_write()`, this argument only applies to the text (e.g. `.txt` or
+#' `.csv`) or spreadsheet file formats (like `.xlsx`). Converting to factors
+#' might be useful for these formats because labelled numeric variables are then
+#' converted into factors and exported as character columns - else, value labels
+#' would be lost and only numeric values are written to the file.
 #' @param verbose Toggle warnings and messages.
 #' @param ... Arguments passed to the related `read_*()` or `write_*()` functions.
 #'
@@ -65,12 +66,13 @@
 #' @section Differences to other packages that read foreign data formats:
 #' `data_read()` is most comparable to `rio::import()`. For data files from
 #' SPSS, SAS or Stata, which support labelled data, variables are converted into
-#' their most appropriate type. The major difference to `rio::import()` is that
-#' `data_read()` automatically converts fully labelled numeric variables into
-#' factors, where imported value labels will be set as factor levels. If a
-#' numeric variable has _no_ value labels or less value labels than values, it
-#' is not converted to factor. In this case, value labels are preserved as
-#' `"labels"` attribute. Character vectors are preserved. Use
+#' their most appropriate type. The major difference to `rio::import()` is for
+#' data files from SPSS, SAS, or Stata, i.e. file types that support
+#' *labelled data*. `data_read()` automatically converts fully labelled numeric
+#' variables into factors, where imported value labels will be set as factor
+#' levels. If a numeric variable has _no_ value labels or less value labels than
+#' values, it is not converted to factor. In this case, value labels are
+#' preserved as `"labels"` attribute. Character vectors are preserved. Use
 #' `convert_factors = FALSE` to remove the automatic conversion of numeric
 #' variables to factors.
 #'
@@ -105,7 +107,7 @@ data_read <- function(path,
     por = .read_spss(path, encoding, convert_factors, verbose, ...),
     dta = .read_stata(path, encoding, convert_factors, verbose, ...),
     sas7bdat = .read_sas(path, path_catalog, encoding, convert_factors, verbose, ...),
-    .read_unknown(path, file_type, convert_factors, verbose, ...)
+    .read_unknown(path, file_type, verbose, ...)
   )
 
   # tell user about empty columns
@@ -188,7 +190,7 @@ data_read <- function(path,
           value_labels <- NULL
           attr(i, "converted_to_factor") <- TRUE
         } else {
-          # else, fall back to numeric
+          # else, fall back to numeric or factor
           i <- as.numeric(i)
         }
 
@@ -288,7 +290,7 @@ data_read <- function(path,
 }
 
 
-.read_unknown <- function(path, file_type, convert_factors, verbose, ...) {
+.read_unknown <- function(path, file_type, verbose, ...) {
   insight::check_if_installed("rio", reason = paste0("to read files of type '", file_type, "'"))
   if (verbose) {
     insight::format_alert("Reading data...")
@@ -317,6 +319,5 @@ data_read <- function(path,
     }
     out <- tmp
   }
-
-  .post_process_imported_data(out, convert_factors, verbose)
+  out
 }