+++ tests/testthat/test-data_read.R
@@ -141,15 +141,38 @@ test_that("data_read - RDS file, matrix, coercible", {
     httr::stop_for_status(request)
     writeBin(httr::content(request, type = "raw"), temp_file)
 
-    expect_message(expect_message(expect_message({
+    expect_message({
       d <- data_read(
         temp_file,
         verbose = TRUE
       )
-    })), regex = "0 out of 5")
+    })
 
     expect_s3_class(d, "data.frame")
     expect_identical(dim(d), c(2L, 5L))
+  })
+})
+
+
+
+# RDS file, preserve class /types -----------------------------------
+
+test_that("data_read - RDS file, preserve class", {
+  withr::with_tempfile("temp_file", fileext = ".rds", code = {
+    request <- httr::GET("https://raw.github.com/easystats/circus/main/data/hiv.rds")
+    httr::stop_for_status(request)
+    writeBin(httr::content(request, type = "raw"), temp_file)
+
+    d <- data_read(temp_file)
+    expect_s3_class(d, "data.frame")
+    expect_identical(
+      sapply(d, class),
+      c(
+        village = "integer", outcome = "integer", distance = "numeric",
+        amount = "numeric", incentive = "integer", age = "integer",
+        hiv2004 = "integer", agecat = "factor"
+      )
+    )
   })
 })